test: Improve TypingSuggestions
